backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
supabase==2.4.1
google-genai==1.3.0
youtube-transcript-api==0.6.2
PyMuPDF==1.24.1
python-multipart>=0.0.12
pydantic==2.6.4
pydantic-settings==2.2.1
sse-starlette>=2.1.0
httpx==0.27.0
nltk==3.8.1
pytest==8.1.1
pytest-asyncio==0.23.6
